+++ CHANGELOG.md
@@ -51,6 +51,8 @@ Remember to align the itemized text with the first line of an item within a list
     behavior (as documented) if the second and third arguments are
     callable, even if other operands are callable as well. See
     [#16413](https://github.com/google/jax/issues/16413).
+  * The deprecated config option `jax_jit_pjit_api_merge`, which did nothing,
+    has been removed.
 
 ## jaxlib 0.4.14